Move D1 loadout/vendors to prefixed folders
