+++ src/app/d1-loadout-builder/GeneratedSet.tsx
