+++ src/app/d1-loadout-builder/LoadoutBuilderItem.tsx
