+++ src/app/d1-loadout-builder/calculate.ts
