+++ src/app/d1-loadout-builder/loadout-builder.scss
