+++ src/app/d1-loadout-builder/routes.ts
