+++ src/app/d1-loadout-builder/utils.ts
@@ -11,7 +11,7 @@ import { D1Item } from '../inventory/item-types';
 
 import _ from 'lodash';
 import { DimStore, D1Store } from '../inventory/store-types';
-import { Vendor } from '../vendors/vendor.service';
+import { Vendor } from '../d1-vendors/vendor.service';
 
 function getBonusType(armorpiece: D1ItemWithNormalStats): string {
   if (!armorpiece.normalStats) {
