+++ src/app/inventory/move-dropped-item.ts
@@ -5,7 +5,7 @@ import { reportException } from '../utils/exceptions';
 import { dimItemService } from './dimItemService.factory';
 import { DimError } from '../bungie-api/bungie-service-helper';
 import { t } from 'app/i18next-t';
-import { PlatformErrorCodes } from '../../../node_modules/bungie-api-ts/user';
+import { PlatformErrorCodes } from 'bungie-api-ts/user';
 import { loadingTracker } from '../shell/loading-tracker';
 import { showNotification } from '../notifications/notifications';
 import { Subject } from 'rxjs';
